Minimization is working now.
Also improved the computation of the fermionic coontribution to use the symmetries of the lattice for an L^3 x L_t
lattice. The modulus of the eigenvalues will be stored to improve the speed.

This improvement will now be implemented in the computation of the gradient and the simultanious computation of the gradient and the function.